Change to floats

+++ api/models/item.py
@@ -5,5 +5,5 @@ from .player import Player
 
 class Item(models.Model):
     player = models.ForeignKey(Player, on_delete=models.CASCADE, editable=False)
-    amount = models.PositiveSmallIntegerField()
+    amount = models.FloatField()
     time_created = models.DateTimeField(auto_now_add=True, editable=False)
